# docs/api/openapi/mesas.openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Mesas API — Artifício RPG"
  description: "API do módulo Mesas. Metadados x-artificio-* definidos por heurística de path."
  version: "0.1.0"
servers:
  - url: https://mesas.artificiorpg.com
    description: Produção
  - url: http://localhost:4000
    description: Desenvolvimento
security: []
paths:
...
  /api/v1/vtt-platforms:
    get:
      operationId: getapi_v1_vtt_platforms
      summary: "Consulta api v1 vtt-platforms"
      x-artificio-owner: mesas
      x-artificio-scope: public
      x-artificio-status: active
      x-artificio-auth: user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/vtt-platforms/admin:
    get:
      operationId: getapi_v1_vtt_platforms_admin
      summary: "Consulta api v1 vtt-platforms admin"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
    post:
      operationId: postapi_v1_vtt_platforms_admin
      summary: "Cria ou executa api v1 vtt-platforms admin"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/vtt-platforms/admin/{id}:
    delete:
      operationId: deleteapi_v1_vtt_platforms_admin_id
      summary: "Remove api v1 vtt-platforms admin id"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
    put:
      operationId: putapi_v1_vtt_platforms_admin_id
      summary: "Substitui api v1 vtt-platforms admin id"
      x-artificio-owner: mesas
      x-artificio-scope: admin
      x-artificio-status: active
      x-artificio-auth: admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
  /api/v1/vtt-platforms/suggest:
    post:
      operationId: postapi_v1_vtt_platforms_suggest
      summary: "Cria ou executa api v1 vtt-platforms suggest"
      x-artificio-owner: mesas
      x-artificio-scope: self-service
      x-artificio-status: active
      x-artificio-auth: user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão
        "500":
          description: Erro interno
...